+++ src/test/java/api/support/http/StorageInterfaceUrls.java
@@ -37,6 +37,10 @@ public class StorageInterfaceUrls {
   public static URL usersStorageUrl(String subPath) {
     return viaOkapiURL("/users" + subPath);
   }
+  public static URL userTenantsStorageUrl(String subPath) {
+    return viaOkapiURL("/user-tenants" + subPath);
+  }
+
   public static URL natureOfContentTermsStorageUrl(String subPath) {
     return viaOkapiURL("/nature-of-content-terms" + subPath);
   }